refactor: move `Classes` into its own file
`widget_style.rs` becomes `widget_style/mod.rs`, with `Classes`,
`HasClasses`, `ClassName` and the class constants in `widget_style/classes.rs`.
All public paths are unchanged.
